Unhoist component back into AppBar
